Start moving stuff to `FreeConvection` submodule
