+++ free_convection/Project.toml
@@ -1,0 +1,12 @@
+name = "FreeConvection"
+uuid = "202ec8e3-8a58-4143-b826-e19fa059e21a"
+authors = ["ali-ramadhan <ali.hh.ramadhan@gmail.com>"]
+version = "0.1.0"
+
+[deps]
+Flux = "587475ba-b771-5e3f-ad9e-33799f191a9c"
+NCDatasets = "85f8d34a-cbdd-5861-8df4-14fed0d494ab"
+OceanParameterizations = "1f034f64-1a41-4bc8-84dc-c8a33b032d87"
+Oceananigans = "9e8cae18-63c1-5223-a75c-80ca9d6e9a09"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
